end of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -50,6 +51,7 @@ namespace AuctionService.Controllers
         }
         
         // create auction
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
@@ -57,7 +59,7 @@ namespace AuctionService.Controllers
             var auction = _mapper.Map<Auction>(auctionDto);
             //todo: add current user as user
             //set seller name
-            auction.Seller="test";
+            auction.Seller= User.Identity.Name;
             //add to db, but not yet commit
             _context.Auctions.Add(auction);
             //map to AuctionDto
@@ -74,6 +76,7 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuctionById), 
                 new {auction.Id}, newAuction);
         }
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -82,8 +85,7 @@ namespace AuctionService.Controllers
             
             if (auction == null) return NotFound();
             
-            //todo: check seller == username
-
+            if (auction.Seller != User.Identity.Name) return Forbid();
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
             auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -98,13 +100,14 @@ namespace AuctionService.Controllers
 
             return BadRequest();
         }
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
             if (auction == null) return NotFound();
 
-            //todo check seller == username
+            if (auction.Seller != User.Identity.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
             //publish to mass transit for remove this auction
